requirements.txt
cassio
langchain
python-dotenv
flask
flask_cors
CLIP @ git+https://github.com/openai/CLIP.git